backend/requirements.txt
fastapi
uvicorn[standard]
python-pptx
pydantic-settings
python-multipart
psycopg[binary]
